inicio da componentização da tela de capa, info e dos jogos de mecânica simples
