+++ src/Components/Game/Info/Info.css
@@ -1,4 +1,10 @@
-.VrumInfoDiv {
+.InfoPageDiv {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.InfoPageDescription {
   margin: 20px 0;
   padding: 0 40px;
   font-weight: 500;
@@ -9,6 +15,6 @@
   overflow-x: hidden;
 }
 
-.VrumInfoDiv::-webkit-scrollbar {
+.InfoPageDescription::-webkit-scrollbar {
   display: none;
 }
