+++ src/Components/Game/SimpleCardGame/SimpleCardGame.css
@@ -1,4 +1,4 @@
-.VrumDiv {
+.SimpleCardGameDiv {
   display: flex;
   flex-direction: column;
   align-items: center;
