+++ src/Components/Game/SimpleCardGame/index.tsx
@@ -1,0 +1,2 @@
+import SimpleCardGame from './SimpleCardGame';
+export default SimpleCardGame;
