+++ src/scenes/OEscolhido/OEscolhido.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import SocketConnection from '../../lib/socket';
 import Background from '../../components/Background';
-import CoverPage from './Cover';
+import CoverPage from '../../components/Game/Cover';
 import InfoPage from './Info';
 import GamePage from './Game';
 import FinishPage from './Finish';
@@ -144,11 +144,12 @@ export default function OEscolhido() {
     case Game.Cover:
       return (
         <CoverPage
+          type="round"
           title={title}
           coverImg={coverImg}
           goBackPage={backToLobby}
-          turnVisibility={turnVisibility}
-          ownerVisibility={ownerVisibility}
+          //turnVisibility={turnVisibility}
+          //ownerVisibility={ownerVisibility}
           infoPage={() => setCurrentGameState(Game.Info)}
           gamePage={() => setCurrentGameState(Game.Game)}
         />
